+++ clinica-medica-comum/src/main/java/br/edu/imepac/comum/repositories/ProntuarioRepository.java
@@ -2,4 +2,4 @@ package br.edu.imepac.comum.repositories;
 import br.edu.imepac.comum.models.Prontuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProntuarioRepository extends JpaRepository<Prontuario, Integer> {}
+public interface ProntuarioRepository extends JpaRepository<Prontuario, Long> {}
